Get data sources / collections implemented in the package

This function returns a table of the data sources / collections that can be downloaded and pre-processed using the getData package

getAvailableDataSources()

Value

a data.frame with the data sources / collections dealt by getData, along with details on each data collection

Details

The output data.frame columns are:

"source": The source acronym

"collection": The collection

"...": ...TODO : finish description

Examples

sources<-getAvailableDataSources()
sources

#>                      source                                       covariate
#> 1                     MODIS                                     Temperature
#> 2                     MODIS                                     Temperature
#> 3                     MODIS                                     Temperature
#> 4                     MODIS                                     Temperature
#> 5                     MODIS                              Vegetation indices
#> 6                     MODIS                              Vegetation indices
#> 7                     MODIS                              Evapotranspiration
#> 8                     MODIS                              Evapotranspiration
#> 9                       GPM                                        Rainfall
#> 10                      GPM                       Rainfall - Night of catch
#> 11                   TAMSAT                                        Rainfall
#> 12                     ERA5                           Wind - Night of catch
#> 13                  MIRIADE Apparent magnitude of the Moon - Night of catch
#> 14                VIIRS DNB               Nighttime lights - Night of catch
#> 15             SRTMGL1_v004             Elevation and elevation derivatives
#> 16               CGLS-LC101                                      Land cover
#> 17                     HRSL                            Built-up, Population
#> 18 WorldPop_100m_Population                                      Population
#> 19             OpenSteetMap                                           Roads
#> 20     Global_Surface_Water                          Open waters / Wetlands
#>                  collection getRemoteData_import_func
#> 1              MOD11A1.v006           getData_modis()
#> 2              MYD11A1.v006           getData_modis()
#> 3              MOD11A2.v006           getData_modis()
#> 4              MYD11A2.v006           getData_modis()
#> 5              MOD13Q1.v006           getData_modis()
#> 6              MYD13Q1.v006           getData_modis()
#> 7              MOD16A2.v006           getData_modis()
#> 8              MYD16A2.v006           getData_modis()
#> 9              GPM_3IMERGDF             getData_gpm()
#> 10             GPM_3IMERGHH             getData_gpm()
#> 11                   TAMSAT          getData_tamsat()
#> 12                     ERA5            getData_era5()
#> 13                  MIRIADE         getData_miriade()
#> 14                VIIRS DNB        getData_viirsdnb()
#> 15             SRTMGL1_v003            getData_srtm()
#> 16               CGLS-LC100            getData_cgls()
#> 17                     HRSL            getData_hrsl()
#> 18 WorldPop_100m_Population        getData_worldpop()
#> 19             OpenSteetMap                          
#> 20     Global_Surface_Water             getData_gsw()
#>    getRemoteData_prepare_func
#> 1         prepareData_modis()
#> 2         prepareData_modis()
#> 3         prepareData_modis()
#> 4         prepareData_modis()
#> 5         prepareData_modis()
#> 6         prepareData_modis()
#> 7         prepareData_modis()
#> 8         prepareData_modis()
#> 9           prepareData_gpm()
#> 10          prepareData_gpm()
#> 11       prepareData_tamsat()
#> 12         prepareData_era5()
#> 13                           
#> 14                           
#> 15                           
#> 16                           
#> 17                           
#> 18                           
#> 19                           
#> 20                           
#>                                                                                         url_metadata
#> 1                                                      https://lpdaac.usgs.gov/products/mod11a1v006/
#> 2                                                      https://lpdaac.usgs.gov/products/myd11a1v006/
#> 3                                                      https://lpdaac.usgs.gov/products/mod11a2v006/
#> 4                                                      https://lpdaac.usgs.gov/products/myd11a2v006/
#> 5                                                      https://lpdaac.usgs.gov/products/mod13q1v006/
#> 6                                                      https://lpdaac.usgs.gov/products/myd13q1v006/
#> 7                                                      https://lpdaac.usgs.gov/products/mod16a2v006/
#> 8                                                      https://lpdaac.usgs.gov/products/myd16a2v006/
#> 9                                        https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDF_06/summary
#> 10                                       https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHH_06/summary
#> 11                                                                   https://www.tamsat.org.uk/about
#> 12     https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=overview
#> 13                                                           http://vo.imcce.fr/webservices/miriade/
#> 14                                      https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html
#> 15                                                     https://lpdaac.usgs.gov/products/srtmgl1v003/
#> 16                                                     https://land.copernicus.eu/global/products/lc
#> 17                                                 https://ciesin.columbia.edu/repository/hrsl/#over
#> 18                         https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0107042
#> 19                                                               https://www.openstreetmap.org/about
#> 20 https://storage.cloud.google.com/global-surface-water/downloads_ancillary/DataUsersGuidev2018.pdf
#>    is_timeSeries
#> 1           TRUE
#> 2           TRUE
#> 3           TRUE
#> 4           TRUE
#> 5           TRUE
#> 6           TRUE
#> 7           TRUE
#> 8           TRUE
#> 9           TRUE
#> 10          TRUE
#> 11          TRUE
#> 12          TRUE
#> 13          TRUE
#> 14          TRUE
#> 15         FALSE
#> 16         FALSE
#> 17         FALSE
#> 18         FALSE
#> 19         FALSE
#> 20         FALSE
#>                                                                                    provider
#> 1                                                                                      NASA
#> 2                                                                                      NASA
#> 3                                                                                      NASA
#> 4                                                                                      NASA
#> 5                                                                                      NASA
#> 6                                                                                      NASA
#> 7                                                                                      NASA
#> 8                                                                                      NASA
#> 9                                                                                      NASA
#> 10                                                                                     NASA
#> 11                                                                                   TAMSAT
#> 12                                                                                Copenicus
#> 13                                                                                    IMCCE
#> 14                                                                                     NOAA
#> 15                                                                                     NASA
#> 16                                                        Copernicus Global Land Operations
#> 17 Facebook Connectivity Lab and Center for International Earth Science Information Network
#> 18                                                                                 WorldPop
#> 19                                                                             OpenSteetMap
#> 20                                                                                      JRC
#>                                                                                                                                     long_name
#> 1                                                           MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 2                                                            MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 3                                                          MODIS/Terra Land Surface Temperature/Emissivity 8-Day L3 Global 1 km SIN Grid V006
#> 4                                                           MODIS/Aqua Land Surface Temperature/Emissivity 8-Day L3 Global 1 km SIN Grid V006
#> 5                                                                          MODIS/Terra Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 6                                                                           MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 7                                                                       MODIS/Terra Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 8                                                                        MODIS/Aqua Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 9                                                                          GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 10                                                                   GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 11                                                    Tropical Applications of Meteorology using SATellite data and ground-based observations
#> 12                                                                                                                                       ERA5
#> 13                                                                            The Virtual Observatory Solar System Object Ephemeris Generator
#> 14                                                                     Visible Infrared Imaging Radiometer Suite (VIIRS) Day/Night Band (DNB)
#> 15                                                                                  NASA Shuttle Radar Topography Mission Global 1 arc second
#> 16                                                                                                           Moderate dynamic land cover 100m
#> 17                                                                                                           High Resolution Settlement Layer
#> 18 Alpha version 2014 estimates of numbers of people per grid square, with national totals adjusted to match UN population division estimates
#> 19                                                                                                                               OpenSteetMap
#> 20                                                                                                                       Global Surface Water
#>    version                                                  DOI
#> 1        6         https://dx.doi.org/10.5067/MODIS/MOD11A1.006
#> 2        6         https://dx.doi.org/10.5067/MODIS/MYD11A1.006
#> 3        6         https://dx.doi.org/10.5067/MODIS/MOD11A2.006
#> 4        6         https://dx.doi.org/10.5067/MODIS/MYD11A2.006
#> 5        6         https://dx.doi.org/10.5067/MODIS/MOD13Q1.006
#> 6        6         https://dx.doi.org/10.5067/MODIS/MYD13Q1.006
#> 7        6         https://dx.doi.org/10.5067/MODIS/MOD16A2.006
#> 8        6         https://dx.doi.org/10.5067/MODIS/MYD16A2.006
#> 9        6           https://doi.org/10.5067/GPM/IMERGDF/DAY/06
#> 10       6           https://doi.org/10.5067/GPM/IMERG/3B-HH/06
#> 11       3                 http://doi.org/10.1038/sdata.2017.63
#> 12       5                                                     
#> 13       2                                                     
#> 14      NA                                                     
#> 15       3 https://dx.doi.org/10.5067/MEASURES/SRTM/SRTMGL1.003
#> 16       2                                                     
#> 17       1                                                     
#> 18       3         10.5258/SOTON/WP00033\n10.5258/SOTON/WP00065
#> 19      NA                                                 <NA>
#> 20       1                                  10.1038/nature20584
#>    spatial_resolution_m temporal_resolution temporal_resolution_unit
#> 1                  1000                   1                      day
#> 2                  1000                   1                      day
#> 3                  1000                   8                      day
#> 4                  1000                   8                      day
#> 5                   250                  16                      day
#> 6                   250                  16                      day
#> 7                   500                   8                      day
#> 8                   500                   8                      day
#> 9                 10000                   1                      day
#> 10                10000                  30                   minute
#> 11                 4000                   1                      day
#> 12                27000                   1                     hour
#> 13                   NA                  NA                     <NA>
#> 14                  450                   1                    month
#> 15                   30                  NA                     <NA>
#> 16                  100                   1                     year
#> 17                   30                  NA                     <NA>
#> 18                  100                  NA                     <NA>
#> 19                   NA                  NA                     <NA>
#> 20                   30                   1                     year
#>                                                                                    spatial_coverage
#> 1                                                                                            Global
#> 2                                                                                            Global
#> 3                                                                                            Global
#> 4                                                                                            Global
#> 5                                                                                            Global
#> 6                                                                                            Global
#> 7                                                                                            Global
#> 8                                                                                            Global
#> 9                                                                                            Global
#> 10                                                                                           Global
#> 11                                                                                           Africa
#> 12                                                                                           Global
#> 13                                                                                           Global
#> 14                                                                                           Global
#> 15                                                                                           Global
#> 16                                                                                           Global
#> 17 Available for some countries : see list here : https://ciesin.columbia.edu/repository/hrsl/#data
#> 18                                                             Africa, Asia, South America, Oceania
#> 19                                                                                           Global
#> 20                                                                                           Global
#>                                                                          url_programmatic_access
#> 1                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/contents.html
#> 2                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD11A1.006/contents.html
#> 3                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A2.006/contents.html
#> 4                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD11A2.006/contents.html
#> 5                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD13Q1.006/contents.html
#> 6                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD13Q1.006/contents.html
#> 7                            https://opendap.cr.usgs.gov/opendap/hyrax/MOD16A2.006/contents.html
#> 8                            https://opendap.cr.usgs.gov/opendap/hyrax/MYD16A2.006/contents.html
#> 9                           https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDF.06/
#> 10                          https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.06/
#> 11                                                        https://www.tamsat.org.uk/data/archive
#> 12               https://dominicroye.github.io/en/2018/access-to-climate-reanalysis-data-from-r/
#> 13                                               http://vo.imcce.fr/webservices/miriade/?ephemcc
#> 14 https://gis.ngdc.noaa.gov/arcgis/rest/services/NPP_VIIRS_DNB/Monthly_AvgRadiance/ImageServer/
#> 15                                  https://e4ftl01.cr.usgs.gov/MEASURES/SRTMGL1.003/2000.02.11/
#> 16                                                                                          <NA>
#> 17                                                                                          <NA>
#> 18                                                                                          <NA>
#> 19                                                                                              
#> 20                                                                                          <NA>
#>                                                                                      url_manual_access
#> 1                                        https://search.earthdata.nasa.gov/search?q=MOD11A1&ok=MOD11A1
#> 2                                        https://search.earthdata.nasa.gov/search?q=MYD11A1&ok=MYD11A1
#> 3                                        https://search.earthdata.nasa.gov/search?q=MOD11A2&ok=MOD11A2
#> 4                                        https://search.earthdata.nasa.gov/search?q=MYD11A2&ok=MYD11A2
#> 5                                        https://search.earthdata.nasa.gov/search?q=MOD13Q1&ok=MOD13Q1
#> 6                                        https://search.earthdata.nasa.gov/search?q=MYD13Q1&ok=MYD13Q1
#> 7                                        https://search.earthdata.nasa.gov/search?q=MOD16A2&ok=MOD16A2
#> 8                                        https://search.earthdata.nasa.gov/search?q=MYD16A2&ok=MYD16A2
#> 9                                           https://search.earthdata.nasa.gov/search?q=GPM_3IMERGDF_06
#> 10                                                                                                    
#> 11                                                        https://www.tamsat.org.uk/data/rfe/index.cgi
#> 12           https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
#> 13                                                       http://vo.imcce.fr/webservices/miriade/?forms
#> 14                                        https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html
#> 15 https://search.earthdata.nasa.gov/search/collection-details?p=C1000000240-LPDAAC_ECS&q=SRTM&ok=SRTM
#> 16                                                                   https://lcviewer.vito.be/download
#> 17                                                   https://ciesin.columbia.edu/repository/hrsl/#data
#> 18                                                      https://www.worldpop.org/geodata/listing?id=16
#> 19                                                                      https://www.openstreetmap.org/
#> 20                                                   https://global-surface-water.appspot.com/download
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           citation
#> 1                                                                                                                                                                                                                                                                                                                                      Wan, Z., Hook, S., Hulley, G. (2015). MOD11A1 MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD11A1.006
#> 2                                                                                                                                                                                                                                                                                                                                       Wan, Z., Hook, S., Hulley, G. (2015). MYD11A1 MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD11A1.006
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> 5                                                                                                                                                                                                                                                                                                                                                                         Didan, K. (2015). MOD13Q1 MODIS/Terra Vegetation Indices 16-Day L3 Global 250m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD13Q1.006
#> 6                                                                                                                                                                                                                                                                                                                                                                          Didan, K. (2015). MYD13Q1 MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD13Q1.006
#> 7                                                                                                                                                                                                                                                                                                                                                  Running, S., Mu, Q., Zhao, M. (2017). MOD16A2 MODIS/Terra Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MOD16A2.006
#> 8                                                                                                                                                                                                                                                                                                                                                   Running, S., Mu, Q., Zhao, M. (2017). MYD16A2 MODIS/Aqua Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006 [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MYD16A2.006
#> 9                                                                                                                                                                                                                      Huffman, G.J., E.F. Stocker, D.T. Bolvin, E.J. Nelkin, Jackson Tan (2019), GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06, Edited by Andrey Savtchenko, Greenbelt, MD, Goddard Earth Sciences Data and Information Services Center (GES DISC), Accessed: [Data Access Date], 10.5067/GPM/IMERGDF/DAY/06
#> 10                                                                                                                                                                                                                                            Huffman, G.J., E.F. Stocker, D.T. Bolvin, E.J. Nelkin, Jackson Tan (2019), GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06, Greenbelt, MD, Goddard Earth Sciences Data and Information Services Center (GES DISC), Accessed: [Data Access Date], 10.5067/GPM/IMERG/3B-HH/06
#> 11 Tarnavsky, E., D. Grimes, R. Maidment, E. Black, R. Allan, M. Stringer, R. Chadwick, F. Kayitakire (2014) Extension of the TAMSAT Satellite-based Rainfall Monitoring over Africa and from 1983 to present Journal of Applied Meteorology and Climate DOI 10.1175/JAMC-D-14-0016.1\nMaidment, R., D. Grimes, R.P.Allan, E. Tarnavsky, M. Stringer, T. Hewison, R. Roebeling and E. Black (2014) The 30 year TAMSAT African Rainfall Climatology And Time series (TARCAT) data set Journal of Geophysical Research DOI: 10.1002/2014JD021927\n\n
#> 12                                                                                                                                                                                                                                                                                      Copernicus Climate Change Service (C3S) (2017): ERA5: Fifth generation of ECMWF atmospheric reanalyses of the global climate . Copernicus Climate Change Service Climate Data Store (CDS), date of access. https://cds.climate.copernicus.eu/cdsapp#!/home
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           This research has made use of IMCCE's Miriade VO tool
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 15                                                                                                                                                                                                                                                                                                                                                                                  NASA JPL (2013). NASA Shuttle Radar Topography Mission Global 1 arc second [Data set]. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MEaSUREs/SRTM/SRTMGL1.003
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 17                                                                                                                                                                                                                                                                                                   Facebook Connectivity Lab and Center for International Earth Science Information Network - CIESIN - Columbia University. 2016. High Resolution Settlement Layer (HRSL). Source imagery for HRSL © 2016 DigitalGlobe. Accessed DAY MONTH YEAR.
#> 18                                                                                                                                                                                            WorldPop (www.worldpop.org - School of Geography and Environmental Science, University of Southampton). 2013. Burkina Faso 100m Population. Alpha version 2010 and 2014 estimates of numbers of people per grid square, with national totals adjusted to match UN population division estimates (http://esa.un.org/wpp/). DOI: 10.5258/SOTON/WP00033
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 20                                                                                                                                                                                                                                                                                                                                            Jean-Francois Pekel, Andrew Cottam, Noel Gorelick, Alan S. Belward, High-resolution mapping of global surface water and its long-term changes. Nature 540, 418-422 (2016). (doi:10.1038/nature20584)