getremotedata: R functions to download & import various spatio-temporal data

This package contains functions to download public domain data, eventually containing spatial and temporal components. Particularly, when web-services are implemented by the data provider, it enables to download the data for a specific region of interest and/or time frame, hence speeding up the overall downloading of data on small regions of interest.

See also