Get data sources / collections implemented in the package

This function returns a table of the data sources / collections that can be downloaded and pre-processed using the getremotedata package

grd_list_collections()

Value

a data.frame with the data sources / collections dealt by getdata, along with details on each data collection

Details

The output data.frame columns are:

"source": The source acronym

"collection": The collection

Examples

sources<-grd_list_collections()
sources

#>        collection source
#> 1            ERA5   ERA5
#> 2         MIRIADE  IMCCE
#> 3     SRTMGL1.003   SRTM
#> 4          TAMSAT TAMSAT
#> 5 VIIRS_DNB_MONTH  VIIRS
#>                                                                                    long_name
#> 1                                     ERA5 hourly data on single levels from 1979 to present
#> 2                            The Virtual Observatory Solar System Object Ephemeris Generator
#> 3 Digital Elevation Model from the NASA Shuttle Radar Topography Mission Global 1 arc second
#> 4    Tropical Applications of Meteorology using SATellite data and ground-based observations
#> 5                     Visible Infrared Imaging Radiometer Suite (VIIRS) Day/Night Band (DNB)
#>                           nature  provider
#> 1                           Wind Copenicus
#> 2 Apparent magnitude of the Moon     IMCCE
#> 3                      Elevation      NASA
#> 4                       Rainfall    TAMSAT
#> 5               Nighttime lights NASA/NOAA
#>                                                                                    url_metadata
#> 1 https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=overview
#> 2                                                       http://vo.imcce.fr/webservices/miriade/
#> 3                                                 https://lpdaac.usgs.gov/products/srtmgl1v003/
#> 4                                                               https://www.tamsat.org.uk/about
#> 5                                        https://eogdata.mines.edu/download_dnb_composites.html
#>                                                    doi version
#> 1                https://doi.org/10.24381/cds.adbb2d47       5
#> 2                                                            2
#> 3 https://dx.doi.org/10.5067/MEASURES/SRTM/SRTMGL1.003       3
#> 4                 http://doi.org/10.1038/sdata.2017.63       3
#> 5            https://doi.org/10.5067/VIIRS/VNP46A1.001      NA
#>   spatial_resolution_m temporal_resolution temporal_resolution_unit
#> 1                27000                   1                     hour
#> 2                   NA                  NA                     <NA>
#> 3                   30                  NA                     <NA>
#> 4                 4000                   1                      day
#> 5                  450                   1                    month
#>   spatial_coverage start_date end_date indicative_latency_days
#> 1           Global         NA       NA                      NA
#> 2           Global         NA       NA                      NA
#> 3           Global         NA       NA                      NA
#> 4           Africa         NA       NA                      NA
#> 5           Global         NA       NA                      NA
#>                                                                                     url_manual_access
#> 1           https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
#> 2                                                       http://vo.imcce.fr/webservices/miriade/?forms
#> 3 https://search.earthdata.nasa.gov/search/collection-details?p=C1000000240-LPDAAC_ECS&q=SRTM&ok=SRTM
#> 4                                                        https://www.tamsat.org.uk/data/rfe/index.cgi
#> 5                                        https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html
#>        status     login
#> 1 Implemented          
#> 2 Implemented          
#> 3 Implemented earthdata
#> 4 Implemented          
#> 5 Implemented          
#>                                                                         url_programmatic_access
#> 1               https://dominicroye.github.io/en/2018/access-to-climate-reanalysis-data-from-r/
#> 2                                               http://vo.imcce.fr/webservices/miriade/?ephemcc
#> 3                           https://opendap.cr.usgs.gov/opendap/hyrax/SRTMGL1.003/contents.html
#> 4                                                        https://www.tamsat.org.uk/data/archive
#> 5 https://gis.ngdc.noaa.gov/arcgis/rest/services/NPP_VIIRS_DNB/Monthly_AvgRadiance/ImageServer/
#>                                                                               crs
#> 1                                                                                
#> 2                                                                                
#> 3 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> 4                                                                                
#> 5                                                                                
#>   param_variables param_roi param_time_range param_credentials
#> 1            TRUE      TRUE             TRUE              TRUE
#> 2           FALSE      TRUE             TRUE             FALSE
#> 3           FALSE      TRUE            FALSE              TRUE
#> 4            TRUE     FALSE             TRUE             FALSE
#> 5            TRUE      TRUE             TRUE             FALSE